fix: Add thread to forward server cond signals to local cond, reducing lock
Previously, the renderer waited on a server-side pthread_cond_t (when shared state was available) protected by a mutex.
Although not heavy, the mutex could lock in renderer thread at the wrong moment—such as when adding buffers
or changing window parameters—delaying the next frame by a few milliseconds and risking a missed vsync.
The thread handling Unix-socket connection handles a lot of other stuff so it can delay sending buffers
from X server to renderer thread, and if X server will ask the renderer to display the not-yet-added buffer
it will block until main thread will actually send this buffer and we will miss vsync.

Now, a separate thread waits on the server condition and forwards its signals to a local pthread_cond_t.
This frees the main thread from blocking on the mutex, reduces missed vsync events (and stuttering).