fix blackscreens on some devices
Fixes #820
For some reasons memfd-based file descriptors can not be passed to processes running in other app sandbox, probably caused by some vendor changes to selinux policies.
Since ASharedMemory was created to work in such conditions it can be used as drop-in replacement for both memfd and ashmem so we can safely pass its file descriptor.